squashed all my commits, sorry for the confusion mate. Update knockout.contextmenu.js
This update will make it work with requirejs and nodejs, as well as standalone.

Update knockout.contextmenu.js

Forgot use strict;

formatting, it's still complaining about a missing newline that is clearly there, hoping it's just a windows error...

I think I figured out what it wants.  This setup is NOT windows friendly lol

karma failing with requirejs for some reason

passing all tests with requirejs code commented out.  The grunt tests fail with it, and I'm not sure why.  Google tells me that grunt doesn't load requirejs and it interfers with the define statement from jasmin, so I'm not sure how to make the tests pass, but the code for requirejs works and my project requires it, pun intended :)